move infinispan data index files
